+++ docs/guides/changelog.md
@@ -7,6 +7,10 @@ description: |-
 
 # Changelog
 
+## 0.1.1 (unreleased)
+
+- Add option to specify nested attributes (YANG containers) to nso_restconf and nso_device_config resources
+
 ## 0.1.0
 
 - Initial release